login mocked, 8 tests green

+++ routes/delUser.ts
@@ -10,14 +10,16 @@ router.delete('/', authMW, async(req: Request, res: Response) =>{
 
     await Pref.deleteMany({userSub: res.locals.sub})
     const [delUser] = await User.find({sub: res.locals.sub})
-    const usedCities = await City.find({_id: {$in: delUser.cities}})
-    await delUser.deleteOne()
-    usedCities.map(async(city) => {
-        const cityUsers = await User.find({cities: {$in: city._id}})
-        if (!cityUsers.length) {
-            await city.deleteOne()
-        }
-    })
+    if (delUser.cities) {
+        const usedCities = await City.find({_id: {$in: delUser.cities}})
+        await delUser.deleteOne()
+        usedCities.map(async(city) => {
+            const cityUsers = await User.find({cities: {$in: city._id}})
+            if (!cityUsers.length) {
+                await city.deleteOne()
+            }
+        })
+    }
     res.sendStatus(200)
 })
